Add FailureBudgeter
Why:
To protect third party resources like authentication servers it would be nice
if we could increase the failure delay in response to a flood of failures.
This would help mitigate some cases like an attacker flooding a broker with
invalid authentication attempts, pressuring some backend server.

This algorithm introduces the idea of a budget of failures per second we
tolerate. With this we can calculate how much of the budget has been consumed.
For example if we allow 5 failures per second and have observed 6 failures
over the last second, then we have overspent by 200ms and would allocate a
200ms delay to the next failure.

The window is short, intended to deal with sustained floods of failures. So
if we don't see any failures for some period (say 5 seconds) we can start a
new window and discard any information about prior failures.